rag_project/requirements.txt
# Requirements for Qdrant RAG System

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Qdrant and embeddings
qdrant-client>=1.7.0
sentence-transformers>=2.2.0

# LangChain for agents
langchain>=0.1.0
langchain-mistralai>=0.1.0
langchain-community>=0.0.20
python-dotenv>=1.0.0

# FastAPI service
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Progress bars
tqdm>=4.65.0

# Optional: For production
python-dotenv>=1.0.0

# Document processing and OCR
opencv-python>=4.8.0
pillow>=10.0.0
pytesseract>=0.3.10
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
easyocr>=1.7.0  # Optional but recommended for fallback OCR
fpdf>=1.7.2  # For PDF generation
